20w25a

+++ auto-editor.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 '''auto-editor.py'''
 
 # external python libraries
@@ -11,15 +12,27 @@ import sys
 import time
 from datetime import timedelta
 from shutil import rmtree
+from operator import itemgetter
 
 # included functions
 from scripts.originalMethod import originalMethod
 from scripts.fastVideo import fastVideo
 
-version = '20w24b hotfix'
+version = '20w25a'
 
 TEMP = '.TEMP'
 CACHE = '.CACHE'
+
+def getFrameRate(path):
+    from re import search
+
+    process = subprocess.Popen(['ffmpeg', '-i', path],
+        stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+    stdout, __ = process.communicate()
+    output = stdout.decode()
+    matchDict = search(r"\s(?P<fps>[\d\.]+?)\stbr", output).groupdict()
+    return float(matchDict["fps"])
+
 
 def file_type(file):
     if(not os.path.isfile(file)):
@@ -63,8 +76,6 @@ if(__name__ == '__main__'):
         help='return what auto-editor thinks the frame rate is.')
     parser.add_argument('--verbose', action='store_true',
         help='display more information when running.')
-    parser.add_argument('--prerun', action='store_true',
-        help='create the cache folder without doing extra work.')
     parser.add_argument('--clear_cache', action='store_true',
         help='delete the cache folder and all of its contents.')
     parser.add_argument('--version', action='store_true',
@@ -103,7 +114,8 @@ if(__name__ == '__main__'):
 
     if(args.clear_cache):
         print('Removing cache')
-        rmtree('.CACHE')
+        if(os.path.isfile('.CACHE')):
+            rmtree('.CACHE')
         if(args.input == []):
             sys.exit()
 
@@ -112,7 +124,6 @@ if(__name__ == '__main__'):
         sys.exit()
 
     INPUT_FILE = args.input[0]
-
     OUTPUT_FILE = args.output_file
     BACK_MUS = args.background_music
     BACK_VOL = args.background_volume
@@ -136,92 +147,124 @@ if(__name__ == '__main__'):
         VIDEO_SPEED = 99999
 
     VERBOSE = args.verbose
-    PRERUN = args.prerun
     HWACCEL = args.hardware_accel
     KEEP_SEP = args.keep_tracks_seperate
 
-    # if input is URL, download as mp4 with youtube-dl
-    if(INPUT_FILE.startswith('http://') or INPUT_FILE.startswith('https://')):
-        print('URL detected, using youtube-dl to download from webpage.')
-        cmd = ["youtube-dl", "-f", "bestvideo[ext=mp4]+bestaudio[ext=m4a]/mp4",
-            INPUT_FILE, "--output", "web_download"]
-        subprocess.call(cmd)
-        print('Finished Download')
-        INPUT_FILE = 'web_download.mp4'
-        OUTPUT_FILE = 'web_download_ALTERED.mp4'
+    if(os.path.isdir(INPUT_FILE)):
+        # get a list of videos sorted by date modified. (date created is too platform dependent)
+        INPUTS = []
+        for filename in os.listdir(INPUT_FILE):
+            if(not filename.startswith('.')):
+                dic = {}
+                dic['file'] = os.path.join(INPUT_FILE, filename)
+                dic['time'] = os.path.getmtime(dic['file'])
+                print(dic['time'])
+                INPUTS.append(dic)
 
-    if(not os.path.isfile(INPUT_FILE)):
-        print('Could not find file:', INPUT_FILE)
-        sys.exit()
+        outputDir = INPUT_FILE+'_ALTERED'
+        newlist = sorted(INPUTS, key=itemgetter('time'), reverse=False)
+
+        INPUTS = []
+        for item in newlist:
+            INPUTS.append(item['file'])
+        del newlist
+
+        # create the new folder for all the outputs
+        try:
+            os.mkdir(outputDir)
+        except OSError:
+            rmtree(outputDir)
+            os.mkdir(outputDir)
+
+    else:
+        outputDir = ''
+        if(os.path.isfile(INPUT_FILE)):
+            # if input is URL, download as mp4 with youtube-dl
+            if(INPUT_FILE.startswith('http://') or INPUT_FILE.startswith('https://')):
+                print('URL detected, using youtube-dl to download from webpage.')
+                cmd = ["youtube-dl", "-f", "bestvideo[ext=mp4]+bestaudio[ext=m4a]/mp4",
+                    INPUT_FILE, "--output", "web_download"]
+                subprocess.call(cmd)
+                print('Finished Download')
+                INPUT_FILE = 'web_download.mp4'
+                OUTPUT_FILE = 'web_download_ALTERED.mp4'
+
+            INPUTS = [INPUT_FILE]
+        else:
+            print('Could not find file:', INPUT_FILE)
+            sys.exit()
 
     if(args.get_auto_fps):
         print(getFrameRate(INPUT_FILE))
         sys.exit()
 
-    isAudio = INPUT_FILE.endswith('.wav') or INPUT_FILE.endswith('.mp3') or INPUT_FILE.endswith('.m4a')
-
-    def getFrameRate(path):
-        from re import search
-
-        process = subprocess.Popen(['ffmpeg', '-i', path],
-            stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-        stdout, __ = process.communicate()
-        output = stdout.decode()
-        matchDict = search(r"\s(?P<fps>[\d\.]+?)\stbr", output).groupdict()
-        return float(matchDict["fps"])
-
-    if(not isAudio):
-        try:
-            frameRate = getFrameRate(INPUT_FILE)
-        except AttributeError:
-            print('Warning! frame rate detection failed.')
-            print('If your video has a variable frame rate, ignore this message.')
-            # convert frame rate to 30 or a user defined value
-            if(GIVEN_FPS is None):
-                frameRate = 30
-            else:
-                frameRate = GIVEN_FPS
-
-            cmd = ['ffmpeg', '-i', INPUT_FILE, '-filter:v', f'fps=fps={frameRate}',
-                TEMP+'/constantVid'+extension, '-hide_banner']
-            if(not VERBOSE):
-                cmd.extend(['-nostats', '-loglevel', '0'])
-            subprocess.call(cmd)
-            INPUT_FILE = TEMP+'/constantVid'+extension
-
     startTime = time.time()
 
-    if(BACK_MUS is None and BACK_VOL != -12):
-        print('Warning! Background volume specified even though no background music was provided.')
+    # all the inputs are stored in the INPUTS variable
+    del INPUT_FILE
 
-    if(KEEP_SEP == False and PRERUN == False and BACK_MUS is None and LOUD_THRESHOLD == 2
-        and NEW_TRAC == None and SILENT_SPEED == 99999 and VIDEO_SPEED == 1
-        and BASE_TRAC == 0 and HWACCEL is None and not isAudio):
+    for INPUT_FILE in INPUTS:
 
-        OUTPUT_FILE = fastVideo(INPUT_FILE, OUTPUT_FILE, SILENT_THRESHOLD,
-            FRAME_SPREADAGE, SAMPLE_RATE, VERBOSE)
-    else:
-        OUTPUT_FILE = originalMethod(INPUT_FILE, OUTPUT_FILE, GIVEN_FPS, FRAME_SPREADAGE,
-            FRAME_QUALITY, SILENT_THRESHOLD, LOUD_THRESHOLD, SAMPLE_RATE, SILENT_SPEED,
-            VIDEO_SPEED, KEEP_SEP, BACK_MUS, BACK_VOL, NEW_TRAC, BASE_TRAC, COMBINE_TRAC,
-            VERBOSE, PRERUN, HWACCEL)
+        dotIndex = INPUT_FILE.rfind('.')
+        extension = INPUT_FILE[dotIndex:]
+        isAudio = extension in ['.wav', '.mp3', '.m4a']
+
+        if(not isAudio):
+            try:
+                frameRate = getFrameRate(INPUT_FILE)
+            except AttributeError:
+                print('Warning! frame rate detection failed.')
+                print('If your video has a variable frame rate, ignore this message.')
+                # convert frame rate to 30 or a user defined value
+                if(GIVEN_FPS is None):
+                    frameRate = 30
+                else:
+                    frameRate = GIVEN_FPS
+
+                cmd = ['ffmpeg', '-i', INPUT_FILE, '-filter:v', f'fps=fps={frameRate}',
+                    TEMP+'/constantVid'+extension, '-hide_banner']
+                if(not VERBOSE):
+                    cmd.extend(['-nostats', '-loglevel', '0'])
+                subprocess.call(cmd)
+                INPUT_FILE = TEMP+'/constantVid'+extension
+
+        if(BACK_MUS is None and BACK_VOL != -12):
+            print('Warning! Background volume specified even though no background music was provided.')
+
+        if(outputDir != ''):
+            newOutput = os.path.join(outputDir, os.path.basename(INPUT_FILE))
+            print(newOutput)
+        else:
+            newOutput = OUTPUT_FILE
+
+        if(KEEP_SEP == False and BACK_MUS is None and LOUD_THRESHOLD == 2
+            and NEW_TRAC == None and SILENT_SPEED == 99999 and VIDEO_SPEED == 1
+            and BASE_TRAC == 0 and HWACCEL is None and not isAudio):
+
+            outFile = fastVideo(INPUT_FILE, newOutput, SILENT_THRESHOLD,
+                FRAME_SPREADAGE, SAMPLE_RATE, VERBOSE)
+        else:
+            outFile = originalMethod(INPUT_FILE, newOutput, GIVEN_FPS, FRAME_SPREADAGE,
+                FRAME_QUALITY, SILENT_THRESHOLD, LOUD_THRESHOLD, SAMPLE_RATE, SILENT_SPEED,
+                VIDEO_SPEED, KEEP_SEP, BACK_MUS, BACK_VOL, NEW_TRAC, BASE_TRAC, COMBINE_TRAC,
+                VERBOSE, HWACCEL)
 
     print('Finished.')
     timeLength = round(time.time() - startTime, 2)
     minutes = timedelta(seconds=round(timeLength))
     print(f'took {timeLength} seconds ({minutes})')
 
-    if(not os.path.isfile(OUTPUT_FILE)):
-        raise IOError(f'Error: The file {OUTPUT_FILE} was not created.')
+    if(not os.path.isfile(outFile)):
+        raise IOError(f'Error: The file {outFile} was not created.')
 
     try:  # should work on Windows
-        os.startfile(OUTPUT_FILE)
+        os.startfile(outFile)
     except AttributeError:
         try:  # should work on MacOS and most Linux versions
-            subprocess.call(['open', OUTPUT_FILE])
+            subprocess.call(['open', outFile])
         except:
             try: # should work on WSL2
-                subprocess.call(['cmd.exe', '/C', 'start', OUTPUT_FILE])
+                subprocess.call(['cmd.exe', '/C', 'start', outFile])
             except:
                 print('Could not open output file.')
